Add ability to download your Taskfile and improve README
