--- src/components/Generator/GeneredTaskfile/Copy/copy.module.css
@@ -1,27 +1,0 @@
-.button {
-	position: sticky;
-	top: 0;
-	float: right;
-	background-color: #10001fc2;
-	color: rgb(255 255 255 / 0.5);
-	border: 0.1rem solid rgb(255 255 255 / 0.5);
-	padding: 1rem;
-	border-radius: 0.2rem;
-	cursor: pointer;
-	width: 5.5rem;
-	font-size: 1rem;
-	transition:
-		color 300ms ease-in-out,
-		border 300ms ease-in-out;
-}
-
-.button:hover {
-	color: #fff;
-	border-color: #fff;
-}
-
-.button.copied,
-.button:hover.copied {
-	color: #2ac93f;
-	border-color: #2ac93f;
-}
