--- src/components/Generator/GeneredTaskfile/Copy/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './Copy';
