+++ src/components/Generator/GeneredTaskfile/SaveFile/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './SaveFile';
